Extract microphone functionality to a separate class
