+++ examples/browser-client-example/public/index.html
@@ -28,6 +28,8 @@
       <button id="initialize">Initialize</button>
       <button id="record">Hold to talk</button>
       <button id="upload">Upload file</button>
+      <button id="close">Close clients</button>
+
       <input type="file" id="fileElem" style="display:none">
       <p id="final"></p>
     </div>
