+++ libraries/browser-client/src/index.ts
@@ -1,4 +1,6 @@
 export * from './speechly'
-export * from './microphone/types'
+export * from './microphone'
 export * from './websocket/types'
 export * from './storage/types'
+export * from './client'
+export * from './audio'
