+++ libraries/browser-client/src/speechly/index.ts
@@ -1,3 +1,2 @@
 export * from './types'
-export { stateToString } from './state'
-export { Client } from './client'
+export * from './segment'
