+++ libraries/browser-client/src/speechly/types.ts
@@ -1,148 +1,3 @@
-import { Microphone } from '../microphone'
-import { Storage } from '../storage'
-import { APIClient } from '../websocket'
-
-/**
- * The options which can be used to configure the client.
- * @public
- */
-export interface ClientOptions {
-  /**
-   * The unique identifier of an app in the dashboard.
-   */
-  appId?: string
-
-  /**
-   * Connect to Speechly upon creating the client instance. Defaults to true.
-   */
-  connect?: boolean
-
-  /**
-   * The unique identifier of a project in the dashboard.
-   */
-  projectId?: string
-
-  /**
-   * @deprecated
-   * The language which is used by the app.
-   */
-  language?: string
-
-  /**
-   * The URL of Speechly login endpoint.
-   */
-  loginUrl?: string
-
-  /**
-   * The URL of Speechly SLU API endpoint.
-   */
-  apiUrl?: string
-
-  /**
-   * The sample rate of the audio to use.
-   */
-  sampleRate?: number
-
-  /**
-   * Whether to output debug statements to the console.
-   */
-  debug?: boolean
-
-  /**
-   * Whether to use auto gain control.
-   * True by default.
-   */
-  autoGainControl?: boolean
-  /**
-   * Whether to output updated segments to the console.
-   */
-  logSegments?: boolean
-
-  /**
-   * Custom microphone implementation.
-   * If not provided, an implementation based on getUserMedia and Web Audio API is used.
-   */
-  microphone?: Microphone
-
-  /**
-   * Custom API client implementation.
-   * If not provided, an implementation based on Speechly SLU WebSocket API is used.
-   */
-  apiClient?: APIClient
-
-  /**
-   * Custom storage implementation.
-   * If not provided, browser's LocalStorage API is used.
-   */
-  storage?: Storage
-}
-
-/**
- * A callback that is invoked whenever the {@link ClientState | client state} changes.
- * @public
- */
-export type StateChangeCallback = (state: ClientState) => void
-
-/**
- * A callback that is invoked whenever current {@link Segment | segment} changes.
- * @public
- */
-export type SegmentChangeCallback = (segment: Segment) => void
-
-/**
- * A callback that is invoked whenever a new tentative transcript is received from the API.
- * @public
- */
-export type TentativeTranscriptCallback = (contextId: string, segmentId: number, words: Word[], text: string) => void
-
-/**
- * A callback that is invoked whenever a new transcript is received from the API.
- * @public
- */
-export type TranscriptCallback = (contextId: string, segmentId: number, word: Word) => void
-
-/**
- * A callback that is invoked whenever new tentative entities are received from the API.
- * @public
- */
-export type TentativeEntitiesCallback = (contextId: string, segmentId: number, entities: Entity[]) => void
-
-/**
- * A callback that is invoked whenever new entity is received from the API.
- * @public
- */
-export type EntityCallback = (contextId: string, segmentId: number, entity: Entity) => void
-
-/**
- * A callback that is invoked whenever new intent (tentative or not) is received from the API.
- * @public
- */
-export type IntentCallback = (contextId: string, segmentId: number, intent: Intent) => void
-
-/**
- * All possible states of a Speechly API client. Failed, NoBrowserSupport and NoAudioConsent states are non-recoverable
- * erroneous states, which should be handled by the end user, according to the semantics of an application.
- * Other states can also be utilized for e.g. enabling and disabling recording buttons or showing the status in the app.
- * It is also possible to use arithmetics for state comparison, e.g. `if (state < speechly.ClientState.Disconnected)`,
- * to react to non-recoverable states.
- * @public
- */
-export enum ClientState {
-  Failed = 0,
-  NoBrowserSupport,
-  NoAudioConsent,
-  __UnrecoverableErrors,
-  Disconnected,
-  Disconnecting,
-  Connecting,
-  Preinitialized,
-  Initializing,
-  Connected,
-  Stopping,
-  Starting,
-  Recording,
-}
-
 /**
  * The smallest component of SLU API, defined by an intent.
  * @public
@@ -255,4 +110,139 @@ export interface Entity {
    * Whether the entity was detected as final.
    */
   isFinal: boolean
+}
+
+/**
+ * A callback that is invoked whenever current {@link Segment | segment} changes.
+ * @public
+ */
+export type SegmentChangeCallback = (segment: Segment) => void
+
+/**
+ * A callback that is invoked whenever a new tentative transcript is received from the API.
+ * @public
+ */
+export type TentativeTranscriptCallback = (contextId: string, segmentId: number, words: Word[], text: string) => void
+
+/**
+ * A callback that is invoked whenever a new transcript is received from the API.
+ * @public
+ */
+export type TranscriptCallback = (contextId: string, segmentId: number, word: Word) => void
+
+/**
+ * A callback that is invoked whenever new tentative entities are received from the API.
+ * @public
+ */
+export type TentativeEntitiesCallback = (contextId: string, segmentId: number, entities: Entity[]) => void
+
+/**
+ * A callback that is invoked whenever new entity is received from the API.
+ * @public
+ */
+export type EntityCallback = (contextId: string, segmentId: number, entity: Entity) => void
+
+/**
+ * A callback that is invoked whenever new intent (tentative or not) is received from the API.
+ * @public
+ */
+export type IntentCallback = (contextId: string, segmentId: number, intent: Intent) => void
+
+/**
+ * Delegate will handle all events received from the Client.
+ * @public
+ */
+export abstract class EventSource {
+  // private stateChangeCb?: StateChangeCallback
+  private segmentChangeCb: SegmentChangeCallback = () => {}
+  private tentativeTranscriptCb: TentativeTranscriptCallback = () => {}
+  private tentativeEntitiesCb: TentativeEntitiesCallback = () => {}
+  private tentativeIntentCb: IntentCallback = () => {}
+  private transcriptCb: TranscriptCallback = () => {}
+  private entityCb: EntityCallback = () => {}
+  private intentCb: IntentCallback = () => {}
+
+  /**
+   * Adds a listener for current segment change events.
+   * @param cb - the callback to invoke on segment change events.
+   */
+  onSegmentChange(cb: SegmentChangeCallback): void {
+    this.segmentChangeCb = cb
+  }
+
+  emitSegmentChange(segment: Segment): void {
+    this.segmentChangeCb(segment)
+  }
+
+  /**
+   * Adds a listener for tentative transcript responses from the API.
+   * @param cb - the callback to invoke on a tentative transcript response.
+   */
+  onTentativeTranscript(cb: TentativeTranscriptCallback): void {
+    this.tentativeTranscriptCb = cb
+  }
+
+  emitTentativeTranscript(contextId: string, segmentId: number, words: Word[], text: string): void {
+    this.tentativeTranscriptCb(contextId, segmentId, words, text)
+  }
+
+  /**
+   * Adds a listener for transcript responses from the API.
+   * @param cb - the callback to invoke on a transcript response.
+   */
+  onTranscript(cb: TranscriptCallback): void {
+    this.transcriptCb = cb
+  }
+
+  emitTranscript(contextId: string, segmentId: number, word: Word): void {
+    this.transcriptCb(contextId, segmentId, word)
+  }
+
+  /**
+   * Adds a listener for tentative entities responses from the API.
+   * @param cb - the callback to invoke on a tentative entities response.
+   */
+  onTentativeEntities(cb: TentativeEntitiesCallback): void {
+    this.tentativeEntitiesCb = cb
+  }
+
+  emitTentativeEntities(contextId: string, segmentId: number, entities: Entity[]): void {
+    this.emitTentativeEntities(contextId, segmentId, entities)
+  }
+
+  /**
+   * Adds a listener for entity responses from the API.
+   * @param cb - the callback to invoke on an entity response.
+   */
+  onEntity(cb: EntityCallback): void {
+    this.entityCb = cb
+  }
+
+  emitEntity(contextId: string, segmentId: number, entity: Entity): void {
+    this.entityCb(contextId, segmentId, entity)
+  }
+
+  /**
+   * Adds a listener for tentative intent responses from the API.
+   * @param cb - the callback to invoke on a tentative intent response.
+   */
+  onTentativeIntent(cb: IntentCallback): void {
+    this.tentativeIntentCb = cb
+  }
+
+  emitTentativeIntent(contextId: string, segmentId: number, intent: Intent): void {
+    this.tentativeIntentCb(contextId, segmentId, intent)
+  }
+
+  /**
+   * Adds a listener for intent responses from the API.
+   * @param cb - the callback to invoke on an intent response.
+   */
+  onIntent(cb: IntentCallback): void {
+    this.intentCb = cb
+  }
+
+  emitIntent(contextId: string, segmentId: number, intent: Intent): void {
+    this.intentCb(contextId, segmentId, intent)
+  }
 }
